# openapi.yaml
...
openapi: 3.0.3
info:
  title: Personal Finance Assistant API
  version: 1.0.0
  description: |
    Comprehensive API for authentication, transaction management,
    split‑bill handling, OCR bill processing, analytics, AI insights,
    and health checks.
servers:
  - url: http://localhost:8080
    description: Development API gateway
paths:
  /auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/register:
    post:
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
...
  /auth/me:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile data
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  role:
                    type: string
  /auth/refresh:
    post:
      summary: Refresh JWT pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /auth/logout:
    post:
      summary: Logout (invalidate refresh token)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /auth/change-password:
    post:
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /upsert/goals:
    post:
      summary: Create a savings goal
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavingsGoalRequest'
      responses:
        '201':
          description: Goal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalResponse'
    get:
      summary: List a user's savings goals
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavingsGoalResponse'
  /upsert/goals/{id}/contribute:
    patch:
      summary: Contribute to a goal
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: amount
          in: query
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalResponse'
  /upsert/goals/{id}:
    delete:
      summary: Delete (archive) a goal
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Goal deleted
  /upsert/budgets:
    post:
      summary: Create a category budget
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryBudgetRequest'
      responses:
        '201':
          description: Budget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetUtilizationResponse'
    get:
      summary: List a user's budgets with utilization
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Budgets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BudgetUtilizationResponse'
  /upsert/budgets/{id}:
    delete:
      summary: Delete a budget
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Budget deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        userId:
          type: string
          format: uuid
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
    RegisterResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...
    SavingsGoalRequest:
      $ref: './upsert-service/dto/SavingsGoalRequest.yaml'
    SavingsGoalResponse:
      $ref: './upsert-service/dto/SavingsGoalResponse.yaml'
    CategoryBudgetRequest:
      $ref: './upsert-service/dto/CategoryBudgetRequest.yaml'
    BudgetUtilizationResponse:
      $ref: './upsert-service/dto/BudgetUtilizationResponse.yaml'
...